requirements.txt
Flask==3.1.0
slack_sdk==3.34.0
slackeventsapi==3.0.3
openai==1.59.7
requests==2.32.3
python-dotenv==1.0.1
numpy==2.2.2
gunicorn==23.0.0